requirements.txt
pdfplumber
python-docx
reportlab
fastapi
uvicorn[standard]
python-multipart
